+++ NEWS.md
@@ -13,6 +13,7 @@
 * Support cut() formatted character columns as well as factor columns in likelihood observation data.
 * Add ``g3_init_val()`` helper, to replace ``gadgetutils::g3_init_guess()``
 * Add ``g3_areas()`` helper to create numbered area vectors
+* Add ``g3_distribution_preview()`` helper to see how observatation data will be converted to array
 
 ## Breaking changes